+++ test/Scripts.Integration.Test/integration-build-project.ps1
@@ -1,6 +1,7 @@
 ﻿param(
     [string] $UnityPath,
     [string] $Platform = "",
+    [string] $UnityVersion = "",
     [Switch] $CheckSymbols
 )
 
@@ -11,13 +12,13 @@ $unityPath = FormatUnityPath $UnityPath
 $buildMethod = BuildMethodFor $Platform
 $outputPath = "$NewProjectBuildPath/$(GetTestAppName $buildMethod)"
 
-Write-Host -NoNewline "Executing ${buildMethod}:"
+Write-Host "Executing ${buildMethod}:"
 $unityArgs = @("-batchmode", "-projectPath ", "$NewProjectPath", "-executeMethod", $buildMethod , "-buildPath", $outputPath, "-quit")
 
 if ($CheckSymbols)
 {
     $symbolServerOutput = RunWithSymbolServer -Callback { RunUnityCustom $unityPath $unityArgs }
-    CheckSymbolServerOutput $buildMethod $symbolServerOutput
+    CheckSymbolServerOutput $buildMethod $symbolServerOutput $UnityVersion
 }
 else
 {