+++ src/Sentry.Unity/Integrations/IApplication.cs
@@ -12,6 +12,7 @@ namespace Sentry.Unity.Integrations
         bool IsEditor { get; }
         string ProductName { get; }
         string Version { get; }
+        string BuildGUID { get; }
         string UnityVersion { get; }
         string PersistentDataPath { get; }
         RuntimePlatform Platform { get; }
@@ -41,6 +42,9 @@ namespace Sentry.Unity.Integrations
         public string ProductName => Application.productName;
 
         public string Version => Application.version;
+
+        public string BuildGUID => Application.buildGUID;
+
         public string UnityVersion => Application.unityVersion;
 
         public string PersistentDataPath => Application.persistentDataPath;