+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Enable native crash support with Mono scripting backend on Android, Windows and Linux ([#751](https://github.com/getsentry/sentry-unity/pull/751))
+- Application-Not-Responding detection ([#771](https://github.com/getsentry/sentry-unity/pull/771))
 
 ## 0.17.0
 
@@ -22,7 +23,6 @@
 - Bump Sentry Java SDK to v6.0.0-rc.1 ([#725](https://github.com/getsentry/sentry-unity/pull/725))
   - [changelog](https://github.com/getsentry/sentry-java/blob/6.0.0-rc.1/CHANGELOG.md?plain=1#L3..L79)
   - [diff](https://github.com/getsentry/sentry-java/compare/6.0.0-alpha.4...6.0.0-rc.1)
-- Application-Not-Responding detection ([#771](https://github.com/getsentry/sentry-unity/pull/771))
 
 ## 0.16.0