+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Enable native crash support with Mono scripting backend on Android, Windows and Linux ([#751](https://github.com/getsentry/sentry-unity/pull/751))
 - Application-Not-Responding detection ([#771](https://github.com/getsentry/sentry-unity/pull/771))
 - Allow uploading sources for debug files ([#773](https://github.com/getsentry/sentry-unity/pull/773))
+- Bump Sentry Java SDK to v6.0.0 ([#787](https://github.com/getsentry/sentry-unity/pull/787))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/6.0.0/CHANGELOG.md?plain=1#L3..L73)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.0.0-rc.1...6.0.0)
 
 ## 0.17.0